move site to blueprints
login and setup is their own blueprints aswell
