+++ templates/modlibrary.html
@@ -97,7 +97,7 @@
       <input type="text" class="form-control invisible" name="modid" id="modid" readonly>
     </div>
     <div class="mb-3">
-      <a href="{{ url_for('newmod') }}"><button type="button" class="btn btn-success tablewidth2">New
+      <a href="{{ url_for('asite.newmod') }}"><button type="button" class="btn btn-success tablewidth2">New
           mod</button></a>
     </div>
     <div class="mb-3">
